+++ environments/homelab/main.tf
@@ -22,6 +22,11 @@ module "vm" {
   datastore_id            = coalesce(each.value.datastore_id, var.default_datastore)
   cloud_init_datastore_id = coalesce(each.value.cloud_init_datastore_id, var.default_cloud_init_datastore)
   boot_disk_interface     = coalesce(each.value.boot_disk_interface, "scsi0")
+  scsi_hardware           = coalesce(each.value.scsi_hardware, var.default_scsi_hardware)
+  disk_iothread           = coalesce(each.value.disk_iothread, var.default_disk_iothread)
+  disk_discard            = coalesce(each.value.disk_discard, var.default_disk_discard)
+  disk_cache              = coalesce(each.value.disk_cache, var.default_disk_cache)
+  disk_file_format        = coalesce(each.value.disk_file_format, var.default_disk_file_format)
   bridge                  = coalesce(each.value.bridge, var.default_bridge)
 
   ipv4         = each.value.ipv4
@@ -33,4 +38,5 @@ module "vm" {
   ssh_public_key      = local.ssh_public_keys[each.key]
   cloud_init_username = coalesce(each.value.cloud_init_username, var.cloud_init_username)
   agent_enabled       = coalesce(each.value.agent_enabled, var.agent_enabled)
+  agent_trim          = coalesce(each.value.agent_trim, var.default_agent_trim)
 }